Update
Added servotimer2 library after timer 1 of normal servo library crashed with timer 1 use of altsoftserial library. The servotimer2 library can be used in the future to generate pwm signals for the rudder and elevator output. An interrupt function is created for the OH SHIT SWITCH so that the pwm signal from the receiver can be measured and switch the SD card measurements on or off during flight. 5 Seperate files are going to be created as the switch is flipped everytime. Each file represents a certain flight stage that must be saved. Still need to test that. 
Furthermore must functionality be added to save rudder and elevator PWM signal as well on D5 and D6. current PCB has no connections and PCB must be redesigned.  PinChangeInt library must be used to measure these PWM signals